tests: move common code to tests/common/mod.rs
Move commonly-used integration_tests code (assert_eq_slices(),
run_test(), and a bunch of consts) out into a common test module so that
other integration tests can use them.

Also, split the run_test() function into keepldr_exec() and
check_output(), which makes the names a little clearer and also means I
don't have to rewrite the whole thing if I want to tweak the Command
args to handle wasmldr tests.
